[Decomp][Chat] Cohesion split: extract message dispatch to ChatMessage.cpp
